Test for navigation from space/index page and removal of factory girl
