--- spec/factories.rb
@@ -1,21 +1,0 @@
-# FactoryGirl.define do
-#   factory :space do
-#     name          'Humpty Dumpty'
-#     description   'nice wall'
-#     price         '19.99'
-#   end
-
-#   factory :space2 do
-#     name          'second space'
-#     description   'second space description'
-#     price         '19.99'
-#   end
-
-#   factory :user do
-#     name          'patrick'
-#   end
-
-# end
-
-
-#we need to make sure this works
